requirements.txt
streamlit
langchain
langchain_openai
pandas
langchain_experimental
openai
tabulate
python-dotenv==1.0.1
geocoder
huggingface_hub
tensorflow 
transformers
mapbox
geopy
streamlit_folium
